% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_correlations.R
\name{add_correlations}
\alias{add_correlations}
\title{Calculate correlation coefficients from MCMC samples of covariance}
\usage{
add_correlations(results_list, hier = FALSE, ngroups = NULL)
}
\arguments{
\item{results_list}{MCMC list of chains, as returned by \code{sample_*} functions}

\item{hier}{(Logical) Whether or not the model is hierarchical}

\item{ngroups}{For hierarchical model, the number of groups}
}
\description{
Calculate correlation coefficients from MCMC samples of covariance
}
