% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_bar_plot.R
\name{bar_plot_matrix}
\alias{bar_plot_matrix}
\title{Draw matrix of correlation bar plots}
\usage{
bar_plot_matrix(dat_lower, dat_upper, param_order = NULL,
  significant_only = TRUE, diag_cex = 2, ...)
}
\arguments{
\item{dat_lower}{Data for lower triangle}

\item{dat_upper}{Data for upper triangle}

\item{param_order}{Character vector of parameter values indicating order}

\item{...}{Additional arguments to \link{single_bar_plot}}
}
\description{
Draw matrix of correlation bar plots
}
