% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_missing.R
\name{bootstrap_missing_hier}
\alias{bootstrap_missing_hier}
\title{Bootstrap missing values in data from fitted multivariate hierarchical model}
\usage{
bootstrap_missing_hier(fit, dat, groups, n = 500, progress = TRUE)
}
\arguments{
\item{fit}{Fitting result, as returned by \code{\link[=fit_mvnorm]{fit_mvnorm()}} or
\code{\link[=fit_mvnorm_hier]{fit_mvnorm_hier()}}. Or, a matrix of samples, as returned by
\code{coda:::as.matrix.mcmc.list(fit)}}

\item{dat}{Data with partially missing values}

\item{groups}{a vector of groups. If \code{NULL}
(default), assume a simple multivariate fit.}

\item{n}{Number of bootstrap samples (default = 500)}

\item{progress}{(Logical) If \code{TRUE} (default), display a progress bar}
}
\value{
If \code{missing_only}, matrix of bootstrapped missing values (\code{nmissing x n}). Otherwise, an array of bootstraps of the full imputed dataset, with
dimensions \code{c(nrow(dat), ncol(dat), n)}.
}
\description{
Bootstrap missing values in data from fitted multivariate hierarchical model
}
\author{
Alexey Shiklomanov
}
