% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_correlations.R
\name{calc_correlations}
\alias{calc_correlations}
\title{Calculate correlation coefficients from MCMC samples of covariance}
\usage{
calc_correlations(sigma_samples, hier = FALSE, ngroups = NULL)
}
\arguments{
\item{hier}{(Logical) Whether or not the model is hierarchical}

\item{ngroups}{For hierarchical model, the number of groups}
}
\description{
Calculate correlation coefficients from MCMC samples of covariance
}
