% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_until_converged.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence of MCMC samples}
\usage{
check_convergence(results_list, model_type, threshold)
}
\arguments{
\item{results_list}{MCMC output of `fit_mvnorm` function}

\item{model_type}{`character(1)`, either `"multi"` or `"hier"`}

\item{threshold}{Threshold for convergence of Gelman diagnostic.}
}
\value{
`logical(1)`, TRUE if converged, FALSE if not or failed to calculate
}
\description{
Check convergence of MCMC samples
}
