% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_slopes.R
\name{cov2rmaslope}
\alias{cov2rmaslope}
\title{Calculate reduced major axis regression slopes from a covariance matrix}
\usage{
cov2rmaslope(m)
}
\arguments{
\item{m}{Variance-covariance matrix}
}
\description{
Calculate reduced major axis regression slopes from a covariance matrix
}
\examples{
cov2rmaslope(cov(iris[, -5]))
}
\author{
Alexey Shiklomanov
}
