% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{default_stickplot}
\alias{default_stickplot}
\title{Stick plot for comparing group means and covariances}
\usage{
default_stickplot(mu_global, Sigma_global, mu_group, Sigma_group, xvar,
  yvar, group_names, par_plot = list(), par_legend = list(),
  xlab = xvar, ylab = yvar, unlog_axes = FALSE, ...)
}
\arguments{
\item{mu_global}{List of global mean statistics}

\item{Sigma_global}{List of global variance-covariance matrix statistics}

\item{mu_group}{List of group mean statistics}

\item{Sigma_group}{List of group variance-covariance matrix statistics}

\item{xvar}{Name of x variable as string (must match row name in \code{matlist})}

\item{yvar}{Name of y variable as string (must match column name in
\code{matlist})}

\item{group_names}{Names of each group. Must match order in \code{mu_group} and
\code{Sigma_group}.}

\item{par_plot}{List of graphical parameters for plot}

\item{par_legend}{List of graphical parameters for legend}

\item{xlab}{Label for x variable}

\item{ylab}{Label for y variable}

\item{unlog_axes}{If TRUE, apply \code{10^x} transformation to axes}

\item{...}{Additional parameters to \code{\link[=prep_stickplot]{prep_stickplot()}}}
}
\description{
Stick plot for comparing group means and covariances
}
