% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ellipse.R
\name{draw_majoraxis}
\alias{draw_majoraxis}
\title{Draw ellipse major axis}
\usage{
draw_majoraxis(ellipse_dat, col = 1, lwd = 1, lty = 1, pch = 19, ...)
}
\arguments{
\item{ellipse_dat}{Ellipse data.frame, as returned by \link{ellipse_axes}}

\item{...}{Additional arguments to \link[graphics:plot]{graphics::plot}}
}
\description{
Draw ellipse major axis
}
