% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{draw_mu}
\alias{draw_mu}
\title{Gibbs draw of mean vector}
\usage{
draw_mu(data, Sigma, mu0, Sigma0)
}
\arguments{
\item{data}{Matrix of data values}

\item{Sigma}{Current draw of covariance matrix}

\item{mu0}{Prior mean values}

\item{Sigma0}{Prior mean covariance}
}
\description{
Gibbs draw of mean vector
}
