% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{draw_Sigma}
\alias{draw_Sigma}
\title{Gibbs draw of covariance matrix}
\usage{
draw_Sigma(data, mu, v0, S0)
}
\arguments{
\item{data}{Matrix of data values}

\item{mu}{Current draw of mean vector}

\item{v0}{Prior Wishart degrees of freedom}

\item{S0}{Prior Wishart rate matrix}
}
\description{
Gibbs draw of covariance matrix
}
