% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{draw_sticks}
\alias{draw_sticks}
\title{Draw ellipse covariance axes from ellipse data}
\usage{
draw_sticks(eps_dat)
}
\arguments{
\item{eps_dat}{Covariance ellipse data frame (see [prep_stickplot()])}
}
\description{
Draw ellipse covariance axes from ellipse data
}
