% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ellipse.R
\name{ellipse_axes}
\alias{ellipse_axes}
\title{Calculate axes for covariance matrix ellipse}
\usage{
ellipse_axes(cov, mean = c(0, 0), prob = 0.95)
}
\arguments{
\item{cov}{Variance-covariance matrix}

\item{mean}{Mean vector}

\item{prob}{Ellipse probability quantile}
}
\description{
Calculate axes for covariance matrix ellipse
}
