% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_bar_plot.R
\name{error_bar_plot}
\alias{error_bar_plot}
\title{Bar plot with error bars}
\usage{
error_bar_plot(height, lower, upper, ...)
}
\arguments{
\item{height}{either a vector or matrix of values describing the
    bars which make up the plot.  If \code{height} is a vector, the
    plot consists of a sequence of rectangular bars with heights
    given by the values in the vector.  If \code{height} is a matrix
    and \code{beside} is \code{FALSE} then each bar of the plot
    corresponds to a column of \code{height}, with the values in the
    column giving the heights of stacked sub-bars making up the
    bar.  If \code{height} is a matrix and \code{beside} is
    \code{TRUE}, then the values in each column are juxtaposed
    rather than stacked.}

\item{lower}{Error bar lower limit}

\item{upper}{Error bar upper limit}

\item{...}{arguments to be passed to/from other methods.  For the
    default method these can include further arguments (such as
    \code{axes}, \code{asp} and \code{main}) and
    \link{graphical parameters} (see \code{\link{par}}) which are passed to
    \code{\link{plot.window}()}, \code{\link{title}()} and
    \code{\link{axis}}.}
}
\description{
Bar plot with error bars
}
