% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mvnorm.R
\name{fit_mvnorm}
\alias{fit_mvnorm}
\title{Fit multivariate normal distribution to data}
\usage{
fit_mvnorm(dat, niter = 5000, priors = list(), inits = list(),
  nchains = 3, autofit = FALSE, max_attempts = 10,
  keep_samples = Inf, threshold = 1.15, save_progress = NULL,
  progress = NULL)
}
\arguments{
\item{dat}{Input data, as a \code{data.frame} or matrix}

\item{niter}{Number of iterations per attempt fitting attempt.}

\item{priors}{Named list of prior overrides. See \code{\link[=gibbs_default_priors]{gibbs_default_priors()}} for
details.}

\item{inits}{Named list of initial conditions}

\item{nchains}{Number of chains (default = 3)}

\item{autofit}{(Logical) If \code{TRUE}, keep re-trying until model has converged (default = \code{FALSE})}

\item{max_attempts}{Maximum number of attempts for \code{autofit} (default = 10)}

\item{keep_samples}{Maximum number of samples to keep in results (default = Inf)}

\item{threshold}{Gelman PSRF threshold for convergence (default = 1.15)}

\item{save_progress}{If not \code{NULL}, path to file in which to save current}

\item{progress}{(Logical) If \code{TRUE}, display a progress bar while sampling
chain status between fitting attempts. If \code{NULL} (default), default to
\code{TRUE} if using a sequential \code{future::plan} and \code{FALSE}}

\item{parallel}{(Logical) Whether or not to run chains in parallel (default = \code{TRUE})}
}
\value{
List containing a tidy summary table of the fitted parameters
(\code{summary_table}), a list of the summary statistics (\code{stats}), and the raw
MCMC samples as an \code{mcmc.list} (\code{samples}).
}
\description{
Fit multivariate normal distribution to data
}
