% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{get_lims}
\alias{get_lims}
\title{Calculate x and y limits from ellipse data}
\usage{
get_lims(eps_dat)
}
\arguments{
\item{eps_dat}{Covariance ellipse data frame (see [prep_stickplot()])}
}
\description{
Calculate x and y limits from ellipse data
}
