% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognorm.R
\name{log_mean}
\alias{log_mean}
\alias{log_var}
\alias{log10_mean}
\alias{log10_var}
\title{Lognormal distribution parameter conversion functions}
\usage{
log_mean(mu, sigma2, base = exp(1))

log_var(mu, sigma2, base = exp(1))

log10_mean(mu, sigma2)

log10_var(mu, sigma2)
}
\arguments{
\item{mu}{Log-scale mean}

\item{sigma2}{Log-scale variance}

\item{base}{Logarithm base. Default is \code{e} (\code{exp(1)})}
}
\description{
Calculate lognormal distribution mean and variance from its log-scale
parameters.
}
