% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerdiag2mat.R
\name{lowerdiag2mat}
\alias{lowerdiag2mat}
\title{Convert a vector of the lower diagonoal of a symmetrical matrix to a matrix}
\usage{
lowerdiag2mat(vec, col_names = TRUE, corr = FALSE, colorder = NULL,
  hier = FALSE)
}
\arguments{
\item{vec}{Vector of lower diagonal values}

\item{col_names}{(Logical) If \code{TRUE} (default), extract row and column names from
vector names formatted as \code{"row..column"}.}

\item{corr}{(Logical) If \code{TRUE}, assume this is a correlation matrix where
the diagonal is fixed at 1 and therefore not stored.}

\item{colorder}{Optional numeric or character vector specifying the desired
column order.}

\item{hier}{(Logical) Whether the vector names also include a group name.
Only used if \code{col_names} is \code{TRUE}.}
}
\description{
Storing just the lower diagonal is an efficient way to store MCMC samples of
a matrix (and, in fact, is how matrices are stored by the samplers in
\code{\link[=fit_mvnorm]{fit_mvnorm()}} and \code{\link[=fit_mvnorm_hier]{fit_mvnorm_hier()}}).
}
