% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_fill_missing.R
\name{mvnorm_fill_missing}
\alias{mvnorm_fill_missing}
\title{Fill missing multivariate data based on known mean vector and covariance
matrix}
\usage{
mvnorm_fill_missing(dat, mu, Sigma, setup = NULL)
}
\arguments{
\item{dat}{Data matrix (observations in rows, variables in columns) with
missing data as NA}

\item{mu}{Mean vector. Must be length \code{ncol(dat)}}

\item{Sigma}{Covariance matrix. Must be dimensions \code{ncol(dat) x ncol(dat)}}

\item{setup}{Information about missingness pattern, as returned by
\code{\link[=setup_missing]{setup_missing()}}. If missing, it will be calculated. When running this
function many times on data with the same missingness pattern, run
\code{\link[=setup_missing]{setup_missing()}} once first to make this calculation more efficient.}
}
\value{
One realization of \code{dat} with missing values imputed
}
\description{
Fill missing multivariate data based on known mean vector and covariance
matrix
}
