% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{prep_ellipse}
\alias{prep_ellipse}
\title{Prepare covariance ellipse for plotting}
\usage{
prep_ellipse(mu, Sigma, xvar, yvar, lty_s = "solid", lty_n = "dashed",
  lwd_s = 2, lwd_n = 0.5, pch_s = 19, pch_n = 4,
  hide_insignificant = FALSE)
}
\arguments{
\item{mu}{List of mean coefficient statistics (see [fit_mvnorm()] and 
[fit_mvnorm_hier()])}

\item{Sigma}{List of variance-covariance matrix statistics (see 
[fit_mvnorm()] and [fit_mvnorm_hier()])}

\item{xvar}{Name of x variable as string (must match row name in `matlist`)}

\item{yvar}{Name of y variable as string (must match column name in `matlist`)}

\item{lty_s}{Line type for significant correlations}

\item{lty_n}{Line type for insignificant correlations}

\item{lwd_s}{Line width for significant correlations}

\item{lwd_n}{Line width for insignificant correlations}

\item{pch_s}{Point shape for significant correlations}

\item{pch_n}{Point shape for insignificant correlations}

\item{hide_insignificant}{If TRUE, set coordinates for insignificant 
correlations to `NA`}
}
\description{
Prepare covariance ellipse for plotting
}
