% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{prep_stickplot}
\alias{prep_stickplot}
\title{Retrieve all covariance ellipse data for plotting}
\usage{
prep_stickplot(mu_global, Sigma_global, mu_group, Sigma_group, xvar, yvar,
  group_names, global_name = "global")
}
\arguments{
\item{mu_global}{List of global mean statistics}

\item{Sigma_global}{List of global variance-covariance matrix statistics}

\item{mu_group}{List of group mean statistics}

\item{Sigma_group}{List of group variance-covariance matrix statistics}

\item{xvar}{Name of x variable as string (must match row name in `matlist`)}

\item{yvar}{Name of y variable as string (must match column name in `matlist`)}

\item{group_names}{Names of each group. Must match order in `mu_group` and 
`Sigma_group`.}
}
\description{
Retrieve all covariance ellipse data for plotting
}
