% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{random_data_hier}
\alias{random_data_hier}
\title{Random hierarchical multivariate data}
\usage{
random_data_hier(mu_global = def_mu_global,
  sigma_global = def_sigma_global, mu_group = def_mu_group,
  sigma_group = def_sigma_group, n = 5000, frac_miss = 0.75,
  all_missing_group = 3, all_missing_col = 2)
}
\arguments{
\item{mu_global}{Global mean vector}

\item{sigma_global}{Global covariance matrix}

\item{mu_group}{Group mean vectors (list)}

\item{sigma_group}{Group covariance matrices (list)}

\item{frac_miss}{Fraction of randomly missing data}

\item{all_missing_group}{Group containing a completely missing column}

\item{all_missing_col}{Column that is missing completely}
}
\description{
Useful for testing.
}
