% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{random_missing}
\alias{random_missing}
\title{Intersperse a random number of NAs through data}
\usage{
random_missing(dat, frac_miss = 0.75)
}
\arguments{
\item{dat}{Data matrix to which to add \code{NA}}

\item{frac_miss}{Fraction of randomly missing data}
}
\description{
Intersperse a random number of NAs through data
}
