% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2mat.R
\name{rearrange_df}
\alias{rearrange_df}
\title{Rearrange result data frame into correct parameter order}
\usage{
rearrange_df(dat_full, params)
}
\arguments{
\item{dat_full}{Summary tibble}

\item{params}{Vector of params, in correct order}
}
\description{
Rearrange result data frame into correct parameter order
}
