% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_until_converged.R
\name{run_until_converged}
\alias{run_until_converged}
\title{Run a sampler until it has converged, or until the maximum number of attempts
has been exceeded}
\usage{
run_until_converged(sampler, model_type, inits, nchains, max_attempts,
  save_progress, threshold, keep_samples, autofit)
}
\arguments{
\item{sampler}{A named list containing \code{fun} (the function to call as the
sampler), \code{init_fun} (the function used to generate the initial
conditions), and \code{args} (a named list of arguments for the sampler
function).}

\item{model_type}{Type of model. Must be either \code{"multi"} or \code{"hier"}.}

\item{inits}{Named list of initial conditions.}

\item{nchains}{Number of chains (default = 3)}

\item{max_attempts}{Maximum number of attempts for \code{autofit} (default = 10)}

\item{save_progress}{If not \code{NULL}, path to file in which to save current}

\item{threshold}{Gelman PSRF threshold for convergence (default = 1.15)}

\item{keep_samples}{Maximum number of samples to keep in results (default = Inf)}

\item{autofit}{(Logical) If \code{TRUE}, keep re-trying until model has converged (default = \code{FALSE})}
}
\value{
List of sample chains. Each sample chain is a named list of sampled parameters.
}
\description{
Run a sampler until it has converged, or until the maximum number of attempts
has been exceeded
}
\author{
Alexey Shiklomanov
}
