% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_fill_missing.R
\name{setup_missing}
\alias{setup_missing}
\title{Analyze missingness pattern in data}
\usage{
setup_missing(dat)
}
\arguments{
\item{dat}{Data matrix (observations in rows, variables in columns) with
missing data as NA}
}
\value{
List describing the missingness pattern in \code{dat} and other relevant
indices for sorting, etc.
}
\description{
Used as an input for \code{\link[=mvnorm_fill_missing]{mvnorm_fill_missing()}}.
}
