% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_bar_plot.R
\name{single_bar_plot}
\alias{single_bar_plot}
\title{Single bar plot panel}
\usage{
single_bar_plot(dat, significant_only = TRUE, ...)
}
\arguments{
\item{dat}{Data for panel}

\item{significant_only}{If `TRUE`, hide columns that are not significantly
difference from zero.}
}
\description{
Single bar plot panel
}
