% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{stick_plot}
\alias{stick_plot}
\title{Stick plot for comparing group means and covariances}
\usage{
stick_plot(mu_global, Sigma_global, mu_group, Sigma_group, xvar, yvar,
  group_names, par_plot = par(), par_legend = par(),
  cols = RColorBrewer::brewer.pal(length(group_names), "Paired"),
  xlab = xvar, ylab = yvar, hide_insignificant_global = FALSE,
  col_g = "black", pch_g = 8, lty_g = "dotted", lwd_g = 3, cex_g = 2,
  hide_insignificant = TRUE, lty_s = "solid", lty_n = "dashed",
  lwd_s = 2, lwd_n = 0.5, pch_s = 19, pch_n = 4)
}
\arguments{
\item{mu_global}{List of global mean statistics}

\item{Sigma_global}{List of global variance-covariance matrix statistics}

\item{mu_group}{List of group mean statistics}

\item{Sigma_group}{List of group variance-covariance matrix statistics}

\item{xvar}{Name of x variable as string (must match row name in `matlist`)}

\item{yvar}{Name of y variable as string (must match column name in `matlist`)}

\item{group_names}{Names of each group. Must match order in `mu_group` and 
`Sigma_group`.}

\item{par_plot}{List of graphical parameters for plot}

\item{par_legend}{List of graphical parameters for legend}

\item{cols}{Colors for group values}

\item{xlab}{Label for x variable}

\item{ylab}{Label for y variable}

\item{hide_insignificant_global}{If TRUE, don't draw global correlation if 
it isn't statistically significant}

\item{col_g}{Color for global values}

\item{pch_g}{Point shape for global correlation}

\item{lty_g}{Line type for global correlation}

\item{lwd_g}{Line width for global correlation}

\item{cex_g}{Point size for global correlation}

\item{hide_insignificant}{If TRUE, set coordinates for insignificant 
correlations to `NA`}

\item{lty_s}{Line type for significant correlations}

\item{lty_n}{Line type for insignificant correlations}

\item{lwd_s}{Line width for significant correlations}

\item{lwd_n}{Line width for insignificant correlations}

\item{pch_s}{Point shape for significant correlations}

\item{pch_n}{Point shape for insignificant correlations}
}
\description{
Stick plot for comparing group means and covariances
}
