% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{summary_barplot}
\alias{summary_barplot}
\title{Column ggplot of summary statistics}
\usage{
summary_barplot(summary_data, varname, aes_list = list(),
  errorbar_list = list(), col_list = list(),
  facet_fun = ggplot2::facet_wrap, facet_list = list(),
  theme_list = list())
}
\arguments{
\item{summary_data}{`tibble` containing summary statistics (e.g. output of `summary_df`)}

\item{varname}{Character string of variable to plot (e.g. "mu", "Sigma", "Corr")}

\item{aes_list}{List of parameters passed to `aes_string`}

\item{errorbar_list}{List of parameters passed to `geom_errorbar`}

\item{col_list}{List of parameters passed to `geom_col`}

\item{facet_fun}{Function for facetting. Default is `facet_wrap`. Use `NULL` to turn off facetting.}

\item{facet_list}{List of parameters passed to `facet_wrap`}

\item{theme_list}{List of parameters passed to `theme`.}
}
\description{
Column ggplot of summary statistics
}
