% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_missing.R
\name{tidy_bootstrap}
\alias{tidy_bootstrap}
\title{Tidy summary table of bootstrapped data}
\usage{
tidy_bootstrap(boot_data, quantiles = c(0.025, 0.975))
}
\arguments{
\item{boot_data}{Bootstrapped data, as returned by \code{\link[=bootstrap_missing]{bootstrap_missing()}} or
\code{\link[=bootstrap_missing_hier]{bootstrap_missing_hier()}}}

\item{quantiles}{Length-2 vector of quantiles to calculate}
}
\value{
\code{data.frame} summarizing the imputed missing values
}
\description{
Tidy summary table of bootstrapped data
}
\author{
Alexey Shiklomanov
}
