% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_missing.R
\name{bootstrap_missing}
\alias{bootstrap_missing}
\title{Bootstrap missing values in data from fitted multivariate model}
\usage{
bootstrap_missing(fit, dat, n = 500)
}
\arguments{
\item{fit}{Fitting result, as returned by \code{\link[=fit_mvnorm]{fit_mvnorm()}} or \code{\link[=fit_mvnorm_hier]{fit_mvnorm_hier()}}.}

\item{dat}{Data with partially missing values}

\item{n}{Number of bootstrap samples (default = 500)}
}
\value{
If \code{missing_only}, matrix of bootstrapped missing values (\code{nmissing x n}). Otherwise, an array of bootstraps of the full imputed dataset, with
dimensions \code{c(nrow(dat), ncol(dat), n)}.
}
\description{
Bootstrap missing values in data from fitted multivariate model
}
\author{
Alexey Shiklomanov
}
