% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{default_stickplot}
\alias{default_stickplot}
\title{Stick plot for comparing group means and covariances}
\usage{
default_stickplot(mu_global, Sigma_global, mu_group, Sigma_group, xvar, yvar,
  group_names, global_name = "global", par_plot = par(),
  par_legend = list(bty = "n"), col_global = "black",
  col_group = RColorBrewer::brewer.pal(length(group_names), "Paired"),
  xlab = xvar, ylab = yvar, pch_g = 8, lty_g = 3, lwd_g = 3,
  cex_g = 2, pch_s = 19, lty_s = 1, lwd_s = 2, cex_s = 1, pch_n = 4,
  lty_n = 2, lwd_n = 0, cex_n = 1)
}
\arguments{
\item{par_plot}{List of graphical parameters for plot}

\item{par_legend}{List of graphical parameters for legend}

\item{xlab}{Label for x variable}

\item{ylab}{Label for y variable}

\item{pch_g}{Point shape for global correlation}

\item{lty_g}{Line type for global correlation}

\item{lwd_g}{Line width for global correlation}

\item{cex_g}{Point size for global correlation}

\item{cols}{Colors for group values}

\item{col_g}{Color for global values}

\item{hide_insignificant_global}{If TRUE, don't draw global correlation if 
it isn't statistically significant}
}
\description{
Stick plot for comparing group means and covariances
}
