% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_ellipse.R
\name{draw_ellipse}
\alias{draw_ellipse}
\title{Draw ellipse}
\usage{
draw_ellipse(ellipse_dat, col = 1, lwd = 1, lty = 1, ...)
}
\arguments{
\item{ellipse_dat}{Ellipse data.frame, as returned by [ellipse_axes]}

\item{...}{Additional arguments to [graphics::plot]}
}
\description{
Draw ellipse
}
