% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{get_ellipse}
\alias{get_ellipse}
\title{Retrieve covariance ellipse for single result}
\usage{
get_ellipse(mu, Sigma, xvar, yvar)
}
\arguments{
\item{mu}{List of mean coefficient statistics (see [fit_mvnorm()] and 
[fit_mvnorm_hier()])}

\item{Sigma}{List of variance-covariance matrix statistics (see 
[fit_mvnorm()] and [fit_mvnorm_hier()])}

\item{xvar}{Name of x variable as string (must match row name in `matlist`)}

\item{yvar}{Name of y variable as string (must match column name in `matlist`)}
}
\description{
Retrieve covariance ellipse for single result
}
