% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_plot.R
\name{prep_stickplot}
\alias{prep_stickplot}
\title{Retrieve all covariance ellipse data for plotting}
\usage{
prep_stickplot(mu_global, Sigma_global, mu_group, Sigma_group, xvar, yvar,
  group_names, global_name = "global", col_global = "black",
  col_group = RColorBrewer::brewer.pal(length(group_names), "Paired"),
  pch_g = 8, lty_g = 3, lwd_g = 3, cex_g = 2, pch_s = 19, lty_s = 1,
  lwd_s = 2, cex_s = 1, pch_n = 4, lty_n = 2, lwd_n = 0, cex_n = 1)
}
\arguments{
\item{mu_global}{List of global mean statistics}

\item{Sigma_global}{List of global variance-covariance matrix statistics}

\item{mu_group}{List of group mean statistics}

\item{Sigma_group}{List of group variance-covariance matrix statistics}

\item{xvar}{Name of x variable as string (must match row name in \code{matlist})}

\item{yvar}{Name of y variable as string (must match column name in
\code{matlist})}

\item{group_names}{Names of each group. Must match order in \code{mu_group} and
\code{Sigma_group}.}

\item{global_name}{Name of global group. Default = "global".}

\item{col_global}{Color for global correlation}

\item{col_group}{List of colors for group correlations}

\item{pch_g}{Point shape for global correlation}

\item{lty_g}{Line type for global correlation}

\item{lwd_g}{Line width for global correlation}

\item{cex_g}{Point size for global correlation}

\item{pch_s}{Point shape for significant group correlations}

\item{lty_s}{Line type for significant group correlations}

\item{lwd_s}{Line width for significant group correlations}

\item{cex_s}{Point size for significant group correlations}

\item{pch_n}{Point shape for insignificant group correlations}

\item{lty_n}{Line type for insignificant group correlations}

\item{lwd_n}{Line width for insignificant group correlations}

\item{cex_n}{Point size for insignificant group correlations}
}
\description{
Retrieve all covariance ellipse data for plotting
}
